refactor(#451): split question form data into multiple files
