+++ designer_v2/lib/features/design/shared/questionnaire/question/models/question_form_data.dart
@@ -1,11 +1,10 @@
-import 'package:flutter/rendering.dart';
 import 'package:studyu_core/core.dart';
 import 'package:studyu_designer_v2/domain/question.dart';
+import 'package:studyu_designer_v2/features/design/shared/questionnaire/question/models/bool_question_form_data.dart';
+import 'package:studyu_designer_v2/features/design/shared/questionnaire/question/models/choice_question_form_data.dart';
+import 'package:studyu_designer_v2/features/design/shared/questionnaire/question/models/scale_question_form_data.dart';
 import 'package:studyu_designer_v2/features/design/shared/questionnaire/question/question_type.dart';
 import 'package:studyu_designer_v2/features/forms/form_data.dart';
-import 'package:studyu_designer_v2/localization/app_translation.dart';
-import 'package:studyu_designer_v2/utils/extensions.dart';
-import 'package:uuid/uuid.dart';
 
 typedef SurveyQuestionFormDataFactory = QuestionFormData Function(
     Question question, List<EligibilityCriterion> eligibilityCriteria);
@@ -120,260 +119,4 @@ abstract class QuestionFormData implements IFormData {
 
   @override
   QuestionFormData copy(); // subclass responsibility
-}
-
-class ChoiceQuestionFormData extends QuestionFormData {
-  ChoiceQuestionFormData({
-    required super.questionId,
-    required super.questionText,
-    required super.questionType,
-    super.questionInfoText,
-    this.isMultipleChoice = false,
-    required this.answerOptions,
-  });
-
-  final bool isMultipleChoice;
-  final List<String> answerOptions;
-
-  @override
-  List<String> get responseOptions => answerOptions;
-
-  factory ChoiceQuestionFormData.fromDomainModel(
-    ChoiceQuestion question,
-    List<EligibilityCriterion> eligibilityCriteria,
-  ) {
-    final data = ChoiceQuestionFormData(
-      questionId: question.id,
-      questionType: SurveyQuestionType.choice,
-      questionText: question.prompt ?? '',
-      questionInfoText: question.rationale ?? '',
-      isMultipleChoice: question.multiple,
-      answerOptions: question.choices.map((choice) => choice.text).toList(),
-    );
-    data.setResponseOptionsValidityFrom(eligibilityCriteria);
-    return data;
-  }
-
-  @override
-  Question toQuestion() {
-    final question = ChoiceQuestion();
-    question.id = questionId;
-    question.prompt = questionText;
-    question.rationale = questionInfoText;
-    question.multiple = isMultipleChoice;
-    question.choices = answerOptions.map(_buildChoiceForValue).toList();
-    return question;
-  }
-
-  @override
-  QuestionFormData copy() {
-    final data = ChoiceQuestionFormData(
-      questionId: const Uuid().v4(),
-      // always regenerate id
-      questionType: questionType,
-      questionText: questionText.withDuplicateLabel(),
-      questionInfoText: questionInfoText,
-      isMultipleChoice: isMultipleChoice,
-      answerOptions: [...answerOptions],
-    );
-    data.responseOptionsValidity = responseOptionsValidity;
-    return data;
-  }
-
-  Choice _buildChoiceForValue(String value) {
-    final choiceId = value.toKey();
-    final choice = Choice(choiceId);
-    choice.text = value;
-    return choice;
-  }
-
-  @override
-  Answer constructAnswerFor(responseOption) {
-    final question = toQuestion() as ChoiceQuestion;
-    final choice = _buildChoiceForValue(responseOption);
-    return question.constructAnswer([choice]);
-  }
-}
-
-class BoolQuestionFormData extends QuestionFormData {
-  BoolQuestionFormData({
-    required super.questionId,
-    required super.questionText,
-    required super.questionType,
-    super.questionInfoText,
-  });
-
-  static Map<String, bool> get kResponseOptions => {
-        tr.form_array_response_options_bool_yes: true,
-        tr.form_array_response_options_bool_no: false,
-      };
-
-  @override
-  List<String> get responseOptions => kResponseOptions.keys.toList();
-
-  factory BoolQuestionFormData.fromDomainModel(
-    BooleanQuestion question,
-    List<EligibilityCriterion> eligibilityCriteria,
-  ) {
-    final data = BoolQuestionFormData(
-      questionId: question.id,
-      questionType: SurveyQuestionType.bool,
-      questionText: question.prompt ?? '',
-      questionInfoText: question.rationale ?? '',
-    );
-    data.setResponseOptionsValidityFrom(eligibilityCriteria);
-    return data;
-  }
-
-  @override
-  Question toQuestion() {
-    final question = BooleanQuestion();
-    question.id = questionId;
-    question.prompt = questionText;
-    question.rationale = questionInfoText;
-    return question;
-  }
-
-  @override
-  BoolQuestionFormData copy() {
-    final data = BoolQuestionFormData(
-      questionId: const Uuid().v4(), // always regenerate id
-      questionType: questionType,
-      questionText: questionText.withDuplicateLabel(),
-      questionInfoText: questionInfoText,
-    );
-    data.responseOptionsValidity = responseOptionsValidity;
-    return data;
-  }
-
-  @override
-  Answer constructAnswerFor(responseOption) {
-    final question = toQuestion() as BooleanQuestion;
-    final value = kResponseOptions[responseOption] as bool;
-    return question.constructAnswer(value);
-  }
-}
-
-class ScaleQuestionFormData extends QuestionFormData {
-  ScaleQuestionFormData({
-    required super.questionId,
-    required super.questionText,
-    required super.questionType,
-    super.questionInfoText,
-    required this.minValue,
-    this.minLabel,
-    required this.maxValue,
-    this.maxLabel,
-    required this.midValues,
-    required this.midLabels,
-    this.initialValue,
-    this.stepSize = 0,
-    this.minColor,
-    this.maxColor,
-  }) : assert(midValues.length == midLabels.length, "midValues.length and midLabels.length must be equal");
-
-  final double minValue;
-  final double maxValue;
-  final String? minLabel;
-  final String? maxLabel;
-  final List<double?> midValues;
-  final List<String?> midLabels;
-  final double stepSize;
-  final double? initialValue;
-  final Color? minColor;
-  final Color? maxColor;
-
-  @override
-  List<double> get responseOptions => toQuestion().values;
-
-  List<Annotation> get midAnnotations {
-    final List<Annotation> midAnnotations = [];
-    for (int i = 0; i < midValues.length; i++) {
-      final value = midValues[i];
-      final label = midLabels[i];
-      if (value != null && label != null && label.isNotEmpty) {
-        final midAnnotation = Annotation()
-          ..value = value.toInt()
-          ..annotation = label;
-        midAnnotations.add(midAnnotation);
-      }
-    }
-    return midAnnotations;
-  }
-
-  factory ScaleQuestionFormData.fromDomainModel(
-    ScaleQuestion question,
-    List<EligibilityCriterion> eligibilityCriteria,
-  ) {
-    final data = ScaleQuestionFormData(
-      questionId: question.id,
-      questionType: SurveyQuestionType.scale,
-      questionText: question.prompt ?? '',
-      questionInfoText: question.rationale ?? '',
-      maxValue: question.maximum,
-      minValue: question.minimum,
-      minLabel: question.minLabel,
-      maxLabel: question.maxLabel,
-      midValues: question.midValues,
-      midLabels: question.midLabels,
-      stepSize: question.step,
-      initialValue: question.initial,
-      minColor: (question.minColor != null) ? Color(question.minColor!) : null,
-      maxColor: (question.maxColor != null) ? Color(question.maxColor!) : null,
-    );
-    data.setResponseOptionsValidityFrom(eligibilityCriteria);
-    return data;
-  }
-
-  @override
-  ScaleQuestion toQuestion() {
-    final question = ScaleQuestion()
-      ..id = questionId
-      ..prompt = questionText
-      ..rationale = questionInfoText
-      ..minimum = minValue
-      ..maximum = maxValue
-      ..step = stepSize
-      ..initial = initialValue
-      ..minColor = minColor?.value
-      ..maxColor = maxColor?.value
-      ..midAnnotations = midAnnotations;
-
-    if (minLabel != null) {
-      question.minLabel = minLabel!;
-    }
-    if (maxLabel != null) {
-      question.maxLabel = maxLabel!;
-    }
-    return question;
-  }
-
-  @override
-  QuestionFormData copy() {
-    final data = ScaleQuestionFormData(
-      questionId: const Uuid().v4(),
-      // always regenerate id
-      questionType: questionType,
-      questionText: questionText.withDuplicateLabel(),
-      questionInfoText: questionInfoText,
-      minValue: minValue,
-      minLabel: minLabel,
-      maxValue: maxValue,
-      maxLabel: maxLabel,
-      stepSize: stepSize,
-      initialValue: initialValue,
-      minColor: minColor,
-      maxColor: maxColor,
-      midLabels: midLabels,
-      midValues: midValues,
-    );
-    data.responseOptionsValidity = responseOptionsValidity;
-    return data;
-  }
-
-  @override
-  Answer constructAnswerFor(responseOption) {
-    final question = toQuestion();
-    return question.constructAnswer(responseOption as double);
-  }
 }
